requirements.txt
chardet>=3.0.4
pyzmq>=17.1.2
